# sandbox-agent/frontend/packages/website/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sandbox-agent",
    "description": "Universal API for automatic coding agents in sandboxes. Supports Claude Code, Codex, OpenCode, and Amp.",
    "contact": {
      "name": "Rivet Gaming, LLC",
      "email": "developer@rivet.gg"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.4.2"
  },
  "servers": [
    {
      "url": "http://localhost:2468"
    }
  ],
  "paths": {
...
    "/v1/acp/{server_id}": {
      "get": {
        "tags": ["v1"],
        "operationId": "get_v1_acp",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Client-defined ACP server id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of ACP envelopes"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ACP server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept SSE responses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["v1"],
        "operationId": "post_v1_acp",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Client-defined ACP server id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agent",
            "in": "query",
            "description": "Agent id required for first POST",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcpEnvelope"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcpEnvelope"
                }
              }
            }
          },
          "202": {
            "description": "JSON-RPC notification accepted"
          },
          "400": {
            "description": "Invalid ACP envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Unknown ACP server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept JSON responses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "ACP server bound to different agent",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "ACP agent process response timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["v1"],
        "operationId": "delete_v1_acp",
        "parameters": [
          {
            "name": "server_id",
            "in": "path",
            "description": "Client-defined ACP server id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "ACP server closed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcpEnvelope": {
        "type": "object",
        "required": ["jsonrpc"],
        "properties": {
          "error": {
            "nullable": true
          },
          "id": {
            "nullable": true
          },
          "jsonrpc": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "params": {
            "nullable": true
          },
          "result": {
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status"],
        "properties": {
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "additionalProperties": {}
      },
...
    }
  },
...
}